pyproject.toml
[tool.poetry]
name = "wordle-with-friends-backend"
version = "0.1.0"
description = ""
authors = ["Daniel Thurau <daniel.n.thurau@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uuid = "^1.30"
mongoengine = "^0.24.1"
Flask = "^2.1.2"
Flask-SocketIO = "^5.1.2"
eventlet = "^0.33.1"
python-socketio = "^5.6.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
